# Avoid deleting a stat if a request raced the reporter