--- a/pkg/activator/handler/concurrency_reporter_test.go
+++ b/pkg/activator/handler/concurrency_reporter_test.go
@@ -460,6 +460,73 @@ func TestConcurrencyReporterHandler(t *testing.T) {
 	}
 }
 
+func TestConcurrencyReporterRace(t *testing.T) {
+	cr, _, cancel := newTestReporter(t)
+	defer cancel()
+
+	base := time.Now()
+
+	// 1. Request in: Creates the stat (first part of handleEvent)
+	eventIn := network.ReqEvent{
+		Time: base,
+		Type: network.ReqIn,
+		Key:  rev1,
+	}
+	stats, _ := cr.getOrCreateStat(eventIn)
+
+	// 2. Report (will remove the stat immediately)
+	cr.report(base.Add(1 * time.Second))
+
+	// 3. HandleEvent (second part of handleEvent)
+	stats.stats.HandleEvent(eventIn)
+
+	// 4. Request out: Creates a new stat, now negative
+	cr.handleEvent(network.ReqEvent{
+		Time: base.Add(2 * time.Second),
+		Type: network.ReqOut,
+		Key:  rev1,
+	})
+
+	// 5. A negative report (if buggy)
+	got := cr.report(base.Add(2 * time.Second))
+	want := []asmetrics.StatMessage{{
+		Key: rev1,
+		Stat: asmetrics.Stat{
+			AverageConcurrentRequests: 1,
+			RequestCount:              1,
+			PodName:                   activatorPodName,
+		},
+	}}
+	if !cmp.Equal(got, want) {
+		t.Error("Unexpected stats (-want +got):", cmp.Diff(want, got))
+	}
+
+	// 6. It continues to work correctly.
+	cr.handleEvent(network.ReqEvent{
+		Time: base.Add(2500 * time.Millisecond),
+		Type: network.ReqIn,
+		Key:  rev1,
+	})
+	cr.handleEvent(network.ReqEvent{
+		Time: base.Add(3 * time.Second),
+		Type: network.ReqOut,
+		Key:  rev1,
+	})
+
+	got = cr.report(base.Add(3 * time.Second))
+	want = []asmetrics.StatMessage{{
+		Key: rev1,
+		Stat: asmetrics.Stat{
+			AverageConcurrentRequests: 0.5,
+			RequestCount:              1,
+			PodName:                   activatorPodName,
+		},
+	}}
+	if !cmp.Equal(got, want) {
+		t.Error("Unexpected stats (-want +got):", cmp.Diff(want, got))
+	}
+}
+
 func TestMetricsReported(t *testing.T) {
 	reset()
 	cr, ctx, cancel := newTestReporter(t)