<!-- Please include the 'why' behind your changes if no issue exists -->
## Proposed Changes

This prevents a race between the report routine and requests flowing in and out. Since we're trying to minimize contention of the request path, the locking routines try to grab as little write-locks as possible, to allow things to progress in parallel.

That breaks though if a report would report AverageConcurrency == 0 and hence marking the stat for deletion. If between this being done and the entry actually being deleted (two separate locks as we only grab a read lock for determining the deletion) comes a nwe request, it'll grab the stat that is now going to be deleted and hence not seen by the next report routine.

The In event is lost and the stats concurrency becomes negative, unrecoverably.

## Benchmark

```
benchmark                                             old ns/op     new ns/op     delta
BenchmarkConcurrencyReporterHandler/sequential-16     263           263           +0.00%
BenchmarkConcurrencyReporterHandler/parallel-16       76.9          82.6          +7.41%

benchmark                                             old allocs     new allocs     delta
BenchmarkConcurrencyReporterHandler/sequential-16     0              0              +0.00%
BenchmarkConcurrencyReporterHandler/parallel-16       0              0              +0.00%

benchmark                                             old bytes     new bytes     delta
BenchmarkConcurrencyReporterHandler/sequential-16     0             0             +0.00%
BenchmarkConcurrencyReporterHandler/parallel-16       0             0             +0.00%
```

Slightly worse as expected, but no biggy imo.

**Release Note**

<!-- Enter your extended release note in the below block. If the PR requires
additional action from users switching to the new release, include the string
"action required". If no release note is required, write "NONE". -->

```release-note
Fixed a bug where the activator's metrics could get stuck and thus scale to and from zero didn't work as expected.
```